+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 
+UNRELEASED -
+------------------
+
+### Added
+- Include both raw and undecoded subject in the message logs.
+
+
 1.0.1 - 2021-01-02
 ------------------